# swagger/v1/swagger.yaml
...
openapi: 3.1.0
info:
  title: Template Ruby Rails Example API
  version: v1
servers:
- url: http://localhost:4000
paths:
...
  "/api/v1/users/me":
    get:
      summary: 내 프로필 조회
      responses:
        '200':
          description: JSON:API document
          content:
            application/vnd.api+json:
              schema:
                "$ref": "#/components/schemas/UserDocument"
        '401':
          description: JSON:API error
          content:
            application/vnd.api+json:
              schema:
                "$ref": "#/components/schemas/ErrorDocument"
        '406':
          description: JSON:API error
          content:
            application/vnd.api+json:
              schema:
                "$ref": "#/components/schemas/ErrorDocument"
        '422':
          description: JSON:API error
          content:
            application/vnd.api+json:
              schema:
                "$ref": "#/components/schemas/ErrorDocument"
        '500':
          description: JSON:API error
          content:
            application/vnd.api+json:
              schema:
                "$ref": "#/components/schemas/ErrorDocument"
      security:
      - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    JsonApiVersion:
      type: object
      required:
      - version
      properties:
        version:
          type: string
          const: '1.1'
...
    ErrorDocument:
      type: object
      required:
      - errors
      - jsonapi
      properties:
        errors:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              status:
                type: string
              code:
                type: string
                enum:
                - NOT_ACCEPTABLE
                - UNSUPPORTED_MEDIA_TYPE
                - INVALID_JSONAPI_DOCUMENT
                - INVALID_QUERY_PARAMETER
                - INVALID_FILTER
                - INVALID_SORT
                - INVALID_INCLUDE
                - INVALID_PAGE
                - RESOURCE_NOT_FOUND
                - RELATIONSHIP_RESOURCE_NOT_FOUND
                - TYPE_MISMATCH
                - ID_MISMATCH
                - CLIENT_GENERATED_ID_UNSUPPORTED
                - RESOURCE_CONFLICT
                - VALIDATION_ERROR
                - INTERNAL_SERVER_ERROR
                - HTTP_ERROR
                - AUTHENTICATION_REQUIRED
                - INVALID_TOKEN
                - TOKEN_EXPIRED
                - USER_INACTIVE
                - EMAIL_ALREADY_REGISTERED
                - INVALID_CREDENTIALS
                - TOKEN_REVOKED
              title:
                type: string
              detail:
                type: string
              source:
                type: object
                properties:
                  pointer:
                    type: string
                  parameter:
                    type: string
                  header:
                    type: string
                required: []
              meta:
                type: object
            required:
            - status
            - code
            - title
            - detail
        jsonapi:
          "$ref": "#/components/schemas/JsonApiVersion"
    UserIdentifier:
      type: object
      required:
      - type
      - id
      properties:
        type:
          type: string
          enum:
          - users
        id:
          type: string
          format: uuid
        meta:
          type: object
    UserAttributes:
      type: object
      required:
      - email
      - isActive
      - createdAt
      - updatedAt
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    UserResource:
      allOf:
      - "$ref": "#/components/schemas/UserIdentifier"
      - type: object
        required:
        - attributes
        - links
        properties:
          attributes:
            "$ref": "#/components/schemas/UserAttributes"
          links:
            type: object
            properties:
              self:
                type: string
            required:
            - self
    UserDocument:
      type: object
      required:
      - data
      - jsonapi
      properties:
        data:
          "$ref": "#/components/schemas/UserResource"
        jsonapi:
          "$ref": "#/components/schemas/JsonApiVersion"
...